Cleanup and migration
Removed uneeded AnimateDiff nodes
Add scaffolding
Modify names
